учтет замечания наставника
