+++ src/view/filters-view.js
@@ -38,5 +38,4 @@ export default class FiltersView extends AbstractView {
   get template() {
     return createFiltersTemplate(this.#filters);
   }
-
 }
