+++ src/view/no-point-view.js
@@ -3,7 +3,7 @@ import AbstractView from '../framework/view/abstract-view.js';
 function createNoPointTemplate() {
   return (
     `<p class="trip-events__msg">
-    Click New Event to create your first point
+      Click New Event to create your first point
     </p>`
   );
 }
